backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "factory-boy>=3.3.3",
    "flake8>=7.3.0",
    "gunicorn>=24.1.1",
    "httpx>=0.28.1",
    "ipython>=9.9.0",
    "isort>=7.0.0",
    "mysqlclient>=2.2.7",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pylint>=4.0.4",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-json-logger>=4.0.0",
    "requests>=2.32.5",
    "shapely>=2.1.2",
]
